Abstracted tooltip to its own action
