+++ src/actions/index.ts
@@ -1,0 +1,1 @@
+export { default as tooltip } from './tooltip';
