+++ src/components/Guitar/subcomponents/Fret/Fret.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
-  import type { Action } from 'svelte/action';
   import type { SelectedNote } from '@/types/SelectedNote';
-  import Tooltip from '@/components/Tooltip';
+  import { tooltip } from '@/actions';
 
   export let note = 'E';
   export let selectedNote: SelectedNote | null = null;
@@ -34,48 +33,6 @@
   $: getIntervalName = (note:string) => (
     getHighlightedNote(note)?.name
   );
-
-  const tooltip: Action<HTMLElement, { text?: string }> = (node, { text }) => {
-    const tooltipElement = new Tooltip({
-      props: { details: text },
-      target: document.body
-    });
-
-    handleResize();
-
-    node.addEventListener('mouseover', handleMouseOver);
-    node.addEventListener('mouseleave', handleMouseLeave);
-    window.addEventListener('resize', handleResize);
-
-    function handleResize() {
-      const { x, y, width } = node.getBoundingClientRect();
-      tooltipElement.$set({
-        x: x + (width / 2),
-        y
-      });
-    }
-
-    function handleMouseOver() {
-      handleResize();
-      tooltipElement.$set({ isVisible: true });
-    }
-
-    function handleMouseLeave() {
-      tooltipElement.$set({ isVisible: false });
-    }
-
-    return {
-      update({ text }) {
-        tooltipElement.$set({ details: text });
-      },
-      destroy() {
-        node.removeEventListener('mouseover', handleMouseOver);
-        node.removeEventListener('mouseleave', handleMouseLeave);
-        window.removeEventListener('resize', handleResize);
-        tooltipElement.$destroy();
-      }
-    };
-  };
 </script>
 
 <button
